+++ realm/realm-library/src/main/cpp/io_realm_internal_UncheckedRow.cpp
@@ -368,7 +368,7 @@ static void finalize_unchecked_row(jlong ptr)
     delete ROW(ptr);
 }
 
-JNIEXPORT jlong JNICALL Java_io_realm_internal_UncheckedRow_nativeGetFinalizerPtr
+JNICALL Java_io_realm_internal_UncheckedRow_nativeGetFinalizerPtr
   (JNIEnv *, jclass)
 {
     TR_ENTER()