+++ realm/realm-library/src/objectServer/java/io/realm/SyncSession.java
@@ -96,15 +96,17 @@ public class SyncSession {
     // This callback will happen on the thread running the Sync Client.
     @KeepMember
     void notifySessionError(int errorCode, String errorMessage) {
-        ObjectServerError error;
+        if (errorHandler == null) {
+            return;
+        }
         ErrorCode errCode = ErrorCode.fromInt(errorCode);
         if (errCode == ErrorCode.CLIENT_RESET) {
-            error = new ClientResetError(errCode, errorMessage);
+            // errorMessage contains the path to the backed up file
+            errorHandler.onClientReset(this, new ClientResetError(errCode, "A Client Reset is required. " +
+                    "Read more here: https://realm.io/docs/realm-object-server/#client-recovery-from-a-backup.",
+                    errorMessage, getConfiguration().getPath()));
         } else {
-            error = new ObjectServerError(errCode, errorMessage);
-        }
-        if (errorHandler != null) {
-            errorHandler.onError(this, error);
+            errorHandler.onError(this, new ObjectServerError(errCode, errorMessage));
         }
     }
 
@@ -134,6 +136,43 @@ public class SyncSession {
          * @param error type of error.
          */
         void onError(SyncSession session, ObjectServerError error);
+
+        /**
+         * An error that indicates the Realm needs to be reset.
+         *
+         * A synced Realm may need to be reset because the Realm Object Server encountered an error and had
+         * to be restored from a backup. If the backup copy of the remote Realm is of an earlier version
+         * than the local copy of the Realm, the server will ask the client to reset the Realm.
+         *
+         * The reset process is as follows: the local copy of the Realm is copied into a recovery directory
+         * for safekeeping, and then deleted from the original location. The next time the Realm for that
+         * URL is opened, the Realm will automatically be re-downloaded from the Realm Object Server, and
+         * can be used as normal.
+         *
+         * Data written to the Realm after the local copy of the Realm diverged from the backup remote copy
+         * will be present in the local recovery copy of the Realm file. The re-downloaded Realm will
+         * initially contain only the data present at the time the Realm was backed up on the server.
+         *
+         * The client reset process can be initiated in one of two ways.
+         *
+         *
+         * The block provided in the
+         * `userInfo` dictionary under `kRLMSyncInitiateClientResetBlockKey` can be called to
+         * initiate the reset process. This block can be called any time after the error is
+         * received, but should only be called if and when your app closes and invalidates every
+         * instance of the offending Realm on all threads (note that autorelease pools may make this
+         * difficult to guarantee).
+         *
+         * If the block is not called, the client reset process will be automatically carried out
+         * the next time the app is launched and the `RLMSyncManager` singleton is accessed.
+         * The value for the `kRLMSyncPathOfRealmBackupCopyKey` key in the `userInfo` dictionary
+         * describes the path of the recovered copy of the Realm. This copy will not actually be
+         * created until the client reset process is initiated.
+         *
+         * @param session {@link SyncSession} this error happened on.
+         * @param error reference to the Client Reset error.
+         */
+        void onClientReset(SyncSession session, ClientResetError error);
     }
 
     String accessToken(final AuthenticationServer authServer) {