+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/ManagementRealmTests.java
@@ -26,6 +26,7 @@ import org.junit.runner.RunWith;
 import java.util.Date;
 import java.util.concurrent.atomic.AtomicReference;
 
+import io.realm.ClientResetError;
 import io.realm.ObjectServerError;
 import io.realm.Realm;
 import io.realm.RealmChangeListener;
@@ -66,6 +67,11 @@ public class ManagementRealmTests extends BaseIntegrationTest {
                     @Override
                     public void onError(SyncSession session, ObjectServerError error) {
                         fail("Realm 1 unexpected error: " + error);
+                    }
+
+                    @Override
+                    public void onClientReset(SyncSession session, ClientResetError error) {
+                        fail("Client Reset");
                     }
                 })
                 .build();